Develop

+++ src/main/java/com/base/basesetup/dto/GstTaxInvoiceDTO.java
@@ -1,5 +1,7 @@
 package com.base.basesetup.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -13,9 +15,9 @@ public class GstTaxInvoiceDTO {
 	private Long id;
 	private String gstChargeAcc;
 	private String gstSubledgerCode;
-	private double gstBdBillAmount;
-	private double gstCrBillAmount;
-	private double gstDbLcAmount;
-	private double gstCrLcAmount;
+	private BigDecimal gstBdBillAmount;
+	private BigDecimal gstCrBillAmount;
+	private BigDecimal gstDbLcAmount;
+	private BigDecimal gstCrLcAmount;
 
 }
